segregate style to new file, move assets
