+++ App.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { View } from 'react-native';
 
-import MenuButton from './components/MenuBottom';
+import MenuButton from './components/MenuBottom/MenuBottom';
 
 export default function App() {
   return (
     <View>
         <MenuButton/>
     </View>
-    
   );
 }
